requirements.txt
Flask==2.3.3
Flask-APScheduler==1.13.1
python-dotenv==1.0.0
requests==2.31.0
gunicorn==20.1.0
